# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Sentinel AI API
  version: 0.1.0
  description: AI Chief Engineer API for deployment risk intelligence.
servers:
  - url: http://localhost:8090
security:
  - bearerAuth: []
...
paths:
...
  /api/deployments/{id}/approval:
    post:
      summary: Approve, block, or request changes for a deployment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated deployment
...
  /api/integration-connections/{provider}/install:
    post:
      summary: Complete simulated or real OAuth installation callback
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [GITHUB, JIRA, CI]
      responses:
        "200":
          description: Connected integration
  /api/integration-connections/{id}/sync:
    post:
      summary: Sync an integration and record health history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Synced integration
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT